backend/requirements.txt
flask>=2.3.0
groq>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.35.0
bcrypt>=4.0.0
python-multipart>=0.0.6
PyJWT>=2.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
requests>=2.31.0 